pyproject.toml
[tool.poetry]
name = "nazo-backend"
version = "0.1.0"
description = ""
authors = ["yanyongyu <https://github.com/yanyongyu>"]
repository = "https://github.com/yanyongyu/nazo-backend"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.1"
uvicorn = "^0.11.3"
pyjwt = "^1.7.1"
python-multipart = "^0.0.5"
sqlalchemy = "^1.3.16"
databases = {extras = ["mysql"], version = "^0.3.1"}
alembic = "^1.4.2"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
